# Don't pass std::string by value to AttributeValueRef()