This causes an accidental copy of the string, and (with VS2017) frees
that string before it's used:
https://github.com/census-instrumentation/opencensus-cpp/issues/22