@bogdandrutu 2018-02-22
Please add benchmarks for the following cases:
* AttributeValueRef("literal long string to prove that we do not copy");
* AttributeValueRef(my_long_string); where my_long_string has 100>chars

Please also add the following cases in the span_test:
* AddAttributes where value is constructed with StrCat;
* AddAttributes where key is constructed with StrCat;
* Maybe replicate these tests for all functions that accept AttributesRef and AddAttribute