@g-easy 2018-02-23
$ bazel run -c opt //opencensus/trace:attribute_value_ref_benchmark

Before:

Benchmark                        Time           CPU Iterations
---------------------------------------------------------------
BM_ConstructFromLiteral          2 ns          2 ns  424733380
BM_ConstructFromString         205 ns        205 ns    3380308

After:

Benchmark                        Time           CPU Iterations
---------------------------------------------------------------
BM_ConstructFromLiteral          2 ns          2 ns  398711129
BM_ConstructFromString           2 ns          2 ns  404455940

:)